Endrer modell for lagring og kafka melding

+++ src/main/kotlin/no/nav/paw/domain/ArbeidssokerRegistrert.kt
@@ -5,7 +5,7 @@ import java.time.ZonedDateTime
 data class ArbeidssokerRegistrert(
     val foedselsnummer: Foedselsnummer,
     val aktorId: AktorId,
-    val registringsId: Long,
+    val registringsId: Int,
     val besvarelse: Besvarelse,
     val opprettetDato: ZonedDateTime
 )
